Lifting State Up on Game Component, Saving history
