+++ react/training-project/src/app/components/Board/styles.scss
@@ -3,7 +3,3 @@
   content: "";
   display: table;
 }
-
-.status {
-  margin-bottom: 10px;
-}
