+++ react/training-project/src/app/components/Game/styles.scss
@@ -6,3 +6,7 @@
 .gameInfo {
   margin-left: 20px;
 }
+
+.status {
+  margin: 0 0 10px;
+}
